Fix crash in start_host when running on a fresh machine
This crash is occurring because of a recent refactoring to share
code between OAuth, Corp, and Cloud provisoning workflows.  It
looks like this problem can only occur if a host has never been
set up and our testing likely occurred on a dirty machine (i.e.
start_host was run with an existing CRD host).

The fix is trivial and highly targeted to this specific codepath.

(cherry picked from commit ec01e2072c6fd69d55a4d99bf187c3796a3911a6)

Bug: b:338602176
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5515192
Commit-Queue: Joe Downing <joedow@chromium.org>
Cr-Original-Commit-Position: refs/heads/main@{#1296182}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5517407
Commit-Queue: Lambros Lambrou <lambroslambrou@chromium.org>
Auto-Submit: Joe Downing <joedow@chromium.org>
Cr-Commit-Position: refs/branch-heads/6422@{#721}
Cr-Branched-From: 9012208d0ce02e0cf0adb9b62558627c356f3278-refs/heads/main@{#1287751}

+++ remoting/host/setup/oauth_host_starter.cc
@@ -89,10 +89,15 @@ void OAuthHostStarter::RemoveOldHostFromDirectory(
     base::OnceClosure on_host_removed) {
   DCHECK_CALLED_ON_VALID_SEQUENCE(sequence_checker_);
   on_host_removed_ = std::move(on_host_removed);
-  directory_service_client_.DeleteHost(
-      *existing_host_id(),
-      base::BindOnce(&OAuthHostStarter::OnDeleteHostResponse,
-                     weak_ptr_factory_.GetWeakPtr()));
+
+  if (existing_host_id().has_value()) {
+    directory_service_client_.DeleteHost(
+        *existing_host_id(),
+        base::BindOnce(&OAuthHostStarter::OnDeleteHostResponse,
+                       weak_ptr_factory_.GetWeakPtr()));
+  } else {
+    std::move(on_host_removed_).Run();
+  }
 }
 
 void OAuthHostStarter::OnRegisterHostResponse(
